Use en_US collator for sorting repo names
This should be configurable but considering we don't internationalize
any other part of the UI, e.g. strings or dates, do the simplest
thing possible for now.

Note that this applies only to repository names, which need to be
sorted manually as we can't trust the filesystem to do so. For other
sorted lists, e.g. tree contents and refs, use the ordering provided
by JGit.

Change-Id: I55dbb9f1ed3b5a20f56f6ea2535ac865c4256196