Add basic blame support
Blame is rendered as a two-cell table, with blame information in the
left cell and the normal blob <pre> in the right.

Currently implemented in a blocking fashion, rendering the entire
blame before sending to the client. Additionally, no state is shared
between requests, since JGit's blame API does not support blaming of a
sequence of history in a way that would make this easier.

Change-Id: I639f1795a9ba6cc0c9677740e41ac34d82c8eb6e