Start versioning Gitiles with build numbers
We do not currently have the notion of a stable release or API; the
largest downstream consumer of Gitiles, googlesource.com, runs
basically at master. Until such time as we do have stable version
numbers, use 0.1 with incremental build numbers as often as we need
to release.

For valid Maven numbering schemes, see:
http://mojo.codehaus.org/versions-maven-plugin/version-rules.html

Change-Id: I732dfbb7ec5c15ff7ee8197e16ce1fb8617e8e5e