Add ArchiveFormat.unregisterAll() method
Users of JGit's archive package are supposed to unregister archive
formats when they're done with them. Since formats in Gitiles are
created/registered in the static initializer of ArchiveFormat, they
will stick around for the lifetime of the Gitiles classloader, but
may still need to be unloaded if a caller knows Gitiles itself is
going to be unloaded.

Change-Id: I4c098f5a21749571f49a49b23615f281886adc3d