Add extra links on hover to blame lines
Issue 5 has some debate about where the blame line link should go. We
were resisting adding 2-3 extra links to every blame line, which would
be a little busy. Instead, use a CSS hover menu. This does raise the
issue of discoverability, but the target audience is someone who is
already interested in clicking on a blame line, so hopefully their
mouse is in the area already.

Change-Id: I6e1e33864cef81f326490d543740ce985249b703