Update patch set 2
Patch Set 2:

I am open to other suggestions that do not involve adding 2 more links to every blame region line.

As far as I'm concerned, this is an argument for the default blame link to be to the diff, since if you hit that you're only one more click away from either of the others.

Patch-set: 2
Label: Verified=0