Update patch set 2
Patch Set 2:

Apparently detecting touch-enabled clients is a no-go because you have things like touchscreen desktops and you can't tell how the user is interacting with them on a moment-by-moment basis.

I think we could have the best of both worlds with a little "v" icon that is hoverable and clickable.

But again, my current feeling is that since the diff and blame are both accessible from the commit link, that is still a sensible and obvious default.

Patch-set: 2