Convert blob view to CodeMirror
CodeMirror is a richer and more performant JavaScript code viewer
than prettify.js. We were hitting alignment and performance issues
with prettify, so it was time to change. Start with the blob box.

Getting CodeMirror running in the dev server is tricky because we are
downloading it from a Maven repo, so we need to build the static
resource zipfile and serve static content from that as well. Borrow
some code from Gerrit to run buck during dev server startup to
prepare that zipfile.

Change-Id: I50452c0d3065a7e804008de0fbdfaee0f523ef37