Use CodeMirror for blame view.
This is much more performant and fixes the alignment issues with
blame regions, but there are a few limitations, either inherent or
for which workarounds were not forthcoming.

Change the hover menu to float on top of the current line rather than
dropping to the next line. The way gutters are implemented in
CodeMirror means there is a new stacking context for each gutter
marker, so we cannot use z-index to float the hover menu on top of
gutter markers in other contexts.

Another current limitation is that the gutter width is fixed, so we
need to hide overflow of long regions. Hopefully the font is small
enough that this is rarely bothersome.

Change-Id: I7fc9864cd8f6aab37f7b0a871fff6b556bfaf756