Enable filetype detection for CodeMirror
Use a stripped-down version of Gerrit's MIME type detector based on
eu.medsea.mimeutil; we don't need some of the subtleties due to the
limited number of MIME types we support.

We need to hard-code the mapping of MIME type to CodeMirror mode so
we can load only the mode JS in the header that is required by the
file in question. This information is actually exposed by the
CodeMirror.mimeModes object, but only when all JS has been loaded. As
a hack, write a script to produce an HTML file that generates the
hard-coded mapping.

The mapping is still somewhat manual for two reasons:
1. CodeMirror's JS has to run in a browser (seriously, I tried
   running it under node.js), so that's never going to happen in the
   Buck build.
2. mimeutil comes up with different MIME types than are registered by
   CodeMirror in a few cases; these are already hard-coded in Gerrit.

Once we have the appropriate mode, it's still currently a little ugly
in the way we pass the codemirror JS/CSS to the header template.
Since we do the MIME type detection only once in BlobSoyData, we end
up having to steal a sub-map from that result and pass it into the
top-level Soy data map so it can be rendered in the header.

Change-Id: I2a5420b274853f0f2d7998377e7831890c94db5f