Allow configuring date formatters with a fixed time zone
By default, the date format used by Gitiles includes the time zone, as
that matches git tools. However, some users/administrators may prefer
normalizing to a particular timezone so times are directly comparable
without doing timezone arithmetic.

Allow this arrangement by setting gitiles.fixedTimeZone to a valid
Java TimeZone ID. This causes all dates in the UI to be implicitly
converted to this timezone, and the now-redundant timezone offset to
be dropped from the output. (Server administrators may communicate to
their users out-of-band about what the fixed timezone is.)

Bug: 48
Change-Id: I11f369471121403d35dae56137fb7fb82a9b10eb