Update patch set 1
Patch Set 1:

> The way you have it now it's only possible for admins to override
 > the default template, not any of the pretty templates. You could
 > tell people to type "?pretty=mySitesCustomFormat" I guess, or we
 > could just list all the available pretty variant names somewhere in
 > the log view. Is that what you were envisioning too? Would that
 > work for you?

Yeah, as a first pass I was figuring this would be something that would just be documented somewhere, and people would have to know to add the appropriate "?pretty=" param to get that formatting. I'm not opposed to making it a more visible option, but maybe as a follow-on CL, after we shake out any issues.

Patch-set: 1