Update patch set 2
Patch Set 2:

Ok, I've thought through this some more (with help from jrn) and I have a plan, which I'd like to run by you. If you want to continue helping to implement that plan I'd of course appreciate it.

So I agree with your idea that we should implement different log pretty formats using variants[1], rather than supporting a fixed set of format names and allowing variants for each name. This is like supporting --pretty=format:"FORMATSTR" in git log, except instead of arbitrary FORMATSTRs, administrators configure a set of formats.

I would like this feature to subsume the current logVariant implementation as
well.

Here's what an example gitiles.config might look like:

 [logFormat "custom1"]
   variant = customLogFormat1

 [logFormat "short"]
   variant = customShortVariant

where CustomLogFormat1 and CustomShortVariant are both defined in MyCustomTemplates.soy.

Out of the box, gitiles will support the built-in formats: "oneline", "short" "medium", "full", "fuller" (maybe not "email"), which will use built-in variants of the same name. If an administrator wants to add their own variant, they can. If it uses the same name as a built-in format but has a different variant name, that's the variant that will get used.

Explicitly listing the variants in gitiles.config also allows us to iterate over them e.g. to generate a list of links for the top of each page.

Follow all that? Sound good?


[1] The current implementation might break down if we ever want log views to include the list of changed files. The reason is, currently, we pass the same parameters to the templates regardless of variant, and we certainly wouldn't want to get the list of changed files if we can know in advance that the variant won't need it.

Patch-set: 2