Update patch set 3
Patch Set 3:

Rebased and resolved conflicts for you.

IIUC your timeline on the Chrome side is you want this within a few days of migration day so a blog post can link to it, is that right? If it looks like this change is going to slip let me know and I'll pick it up.

Patch-set: 3