Update patch set 1
Patch Set 1:

The case Chromium was complaining about was a file that was modified once 130k commits or so ago. On my workstation, the findLastCommit step takes <4s on C git and <7s on JGit, hence my claim that an optimized implementation is likely not worth it. Contrast this with googlesource.com where it's 15-20s or more.

Patch-set: 1
Label: Verified=0