Update patch set 1
Patch Set 1:

> Per IM, (start, count) would be more consistent with 'git blame
 > --porcelain and 'diff -u'. Thoughts?

Will do.

 > * It would be handy to have committer info, since the committer
 > date seems to be what chromium people pay attention to.

I don't think they have a strong preference except that they want log output to be appear monotonic. This is not an issue with blame (and they haven't complained about this in the web UI).

 > * It might also be nice to omit the author info after the first
 > time a commit is mentioned, for a shorter payload.

As mentioned elsewhere I think interleaving is ugly but would be ok with a separate section.

I'm skeptical that it would be noticeably shorter when zlib compressed anyway.

 > * The corresponding start line for a region in the source commit
 > would be handy e.g. for making hyperlinks to the right line.

Yes, we should add this later.

Patch-set: 1