Update patch set 3
Patch Set 3:

For context, this is needed because it's currently impossible to construct an identical commit object from the JSON api (particularly because of the way that the author/committer times are handled. Instead of trying to fix that, I thought it would be better to just dump the actual /real/ commit object text via the TEXT api. Now format=TEXT can be used to reproduce the real objects that they're formatting (blob and tree already support it, though I think there may be a bug if a path has a newline in it for the tree view... but who does that anyway?)

Patch-set: 3
Label: Verified=0