Update patch set 3
Patch Set 3:

> (particularly because of the way that the author/committer times
 > are handled.

This is primarily because format=JSON respects the gitiles.fixedTimeZone option, right? I wonder if we should just fix that as well.

That said I have no objections to this functionality. There are other things we don't handle properly right now, like encoding, that might prevent this.

Patch-set: 3