Update patch set 6
Patch Set 6:

So now the html links are also working. However to to so, I needed to have Map<String, Object> instead of Map<String, String> to tell the soy framework we really want to pass in unescaped html, because we need the 
UnsafeSanitizedContentOrdainer from Soy to put Strings into. This makes the rest of the code look ugly. (Lots of toString() and (Object)casts).

Patch-set: 6
Label: Verified=0