Render markdown at /+doc/
Parse and render any file ending with ".md" as markdown using the
pegdown processor.  This allows a repository to be used as a
documentation source.

If "navbar.md" is found in the top level directory it is rendered as a
horizontal bar across the top of the page.  A navbar.md structure is a
simple outline:

  * [Home](/index.md)
  * [APIs](/api/index.md)
  * [Source](/src/main/java/index.md)

Links starting with "/" and ending in ".md" are resolved as
relative to the top of the Git repository.

Change-Id: Ic111628c59cfadfdca37bf0cc637ee8a14d54c37