Recognize special note/aside blocks in markdown
Extend markdown syntax with a special aside format useful for
documentation writing.

  *** note
  Pay special attention here!

  You may have to watch out for crazy developers.
  ***

A note block starts with *** on its own line and ends with ***\n on
its own line. Everything inside of the block is read as Markdown and
reprocessed through a recursive call whose AST is inlined.

Three kinds of notes are recognized, which are given different CSS:

  *** note
  *** promo
  *** aside

Change-Id: Icbfd27d3afd7d0137e37b7d5e821938c70207cf1