Markdown: Support *** note extensions inside |||---|||
By adding the extensions to the recursive processor it is now possible
to use the note/aside/promo blocks inside of the cols extension.  This
can help draw attention to the content of a specific column:

  |||---|||
  ## Benefits
  Fast.

  Simple.

  *** note
  ## Drawbacks
  Nesting is hard.
  ***
  |||---|||

The ColsSerializer now recognizes the inner DivNode used by
the note block and starts a new column for its contents.

Change-Id: Id0f00af1f9a9e11a86c40505651c661283c471cf