Markdown: apply prettify to verbatim blocks
Verbatim blocks may optionally specify a language, e.g.:

  ```java
  import java.io.InputStream;
  ```

If the named language is known to be supported by prettify parser
try to format with syntax highlighting and include that styled
result in the HTML.

Use a cache to avoid recreating the Prettify instance each time
a block needs to be formatted. Prettify can be run over multiple
inputs but is not thread-safe due to updates of file extension
maps while loading a language that has not been processed before.

Change-Id: Ia37628d7c440df0170a21e69383ab69b20e85beb