Markdown: configure optional banner in navbar.md
The navbar.md file can now define a documentation set header and
optional project logo:

  # Gerrit Code Review
  [logo]: http://storage.googleapis.com/gerrit-static/diffy-w200.png
  [home]: /index.md

  * [Home][home]
  * [APIs](/api/index.md)
  * [Source](/src/main/java/index.md)

The optional site header must be an H1 header and must be specified to
use the logo or home references.

The optional project logo is specified by a reference named "logo".
As images (or any other file type) are not served directly from
Gitiles this must be a link to an external service.  Recommended size
is no taller than 44px.

The optional "home" reference will create an <a href> to wrap the logo
and H1 inside an anchor to jump back to the logical top of the
documentation set. It can also be used inside the outline part of the
navbar to create a "Home" link.

These nodes are extracted from navbar.md and hoisted into the top of
the Soy template above the navbar's usual rendering.

   +-------+
   | Diffy |                       |
   |       |   Gerrit Code Review  |  APIs
   | Logo  |                       |
   +-------+

   Home    APIs    Source

The current page's H1 title is extracted and displayed to the right,
e.g. "APIs" above.

Change-Id: I70b4fd5dd51d4a92d328692740dd6cfa9655e84c