Update patch set 1
Patch Set 1:

Hm. On second thought maybe not for text? Insofar as there is only a streaming interface for text.

We don't currently serve text results for log and blame, but I can see if we did that we would want to stream them in the same way we do for HTML.

Patch-set: 1
Label: Code-Review=0