RFC Markdown: render /+/foo.md as markdown
Default to the rendered version of a markdown file when it is
clicked on in a tree or navigated to using /+/ operator in URL.

Viewing on /+/ feels messy. Should this be a redirect to /+doc/?

Building URLs with /+show/ is not available in GitilesView but it can
be parsed out of the URL.  Adding Type.SHOW is messy as there is no
servlet for this because its handled by the path servlet.

Change-Id: I385e6fb6fe2cfcddf6ef8b0d0f4192e920e848e5