Update patch set 3
Patch Set 3: Code-Review-1

(9 comments)

This is a great start, thanks for your help on this.

However, this is just the "easy" part. The trickier part is passing a stateless authentication token in case the repo contents are not world-readable, which may happen on a per-branch level.

In an environment like Gerrit Code Review, the Repository object attached to the request might:

* have repo-level access controls, which would have already been verified by the RepositoryFilter
* have ref-level access controls, which you can't get at explicitly, but are implicit in the list of refs returned by the Repository instance (see FilteredRepository.java in the gerrit gitiles plugin for an example)

In terms of structuring a series of changes, I think something like:

* Add a method to GitilesAccess like boolean isRepositoryWorldReadable() (not totally happy with that name, but you get the idea.
* This change, plus a stub to return 403 if isRepositoryWorldReadable() returns false.
* A change to implement token-passing for non-world-readable repositories. Contact me privately and I can share our design for how the authentication tokens will work.

Patch-set: 3
Reviewer: Dave Borowitz <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1