Update patch set 1
Patch Set 1:

> Does something like this belong in JGit as well? What does git log
 > --oneline do with these commits?

git-core's `git log --oneline` does the same thing as JGit, it shows the entire message with s/\n/ /g. Its horrible.

I think JGit is correct in its behavior in RevCommit.getShortMessage(), and this is a gitiles specific choice to display less.

FWIW gitk works exactly the way this gitiles CL does, it shows up to the first LF. But that formatting choice is local to gitk.

Patch-set: 1