Add ROOTED_DOC view to serve at Markdown at root of a host
This is to support serving [1] at https://gerritcodereview.com/ without
having the messy prefix of "homepage/+doc/HEAD" in the path. Google is
hosting this DNS name and will run a server installing RootedDocServlet
at "/" for gerritcodereview.com. This will mount and serve Markdown,
making a new homepage for Gerrit.

[1] https://gerrit.googlesource.com/homepage/+doc/HEAD/

Change-Id: I73ae4451586a5acbb9187b3c4dd13a05e88a1926