RFC RawServlet to dangerously serve user content
Simple implementation to replace the safe PathServlet with a
dangerous version that spits back the exact bytes stored in Git.
This may be HTML, JavaScript or a Flash content. It should never
be run on the same domain as an authenticated Gitiles instance.

Currently punts on setting up content type for the files stored.
A file extension to MIME type mapping will be required.

Change-Id: Ib6984694e83aef0649d8ad02b0a0fcc8181725e7