BlameCacheImpl: Don't store Repository reference in key
Storing the reference in the key means we need to remember to clear
it in all code paths. In this case there is only a single load code
path and we are clearing the reference, but in the case of something
like Gerrit's H2CacheImpl, we might unintentionally hold on to the
reference for too long. We now consider this an antipattern and
prefer using the non-LoadingCache interface so we don't accidentally
leak transient data in cache keys.

Change-Id: Ic661829d89e346a2d808ddfcf452a65db98c4552