Markdown: Disabled URI value filter for URLs
By default, HtmlBuilder.anchorHref considers URLs with "/../" dangerous
and returns #zSoyz instead. This lead to some code complexity in
MarkdownToHtml.href: we had to resolve a path ourselves (normally done
by browser). It also inconvenient for the user when they see #zSoyz without any
explanation.

Initially the rationale for disallowing "/../" in FilterNormalizeUri was to
prevent a malicious user to inject "../../" in $local parameter. We
filter out scripts, so this case does not apply to Markdown.

This CL disables the check by FitlerNormalizeUri.valueFilter and
simplifies MMarkdownToHtml.href

Change-Id: Iacf664dc4034d1a5c4707f48b9e121f742aaa8ce