Markdown: Disabled URI value filter
By default, HtmlBuilder.anchorHref considers URLs with "/../" dangerous
and returns #zSoyz instead. This lead to some code complexity in
MarkdownToHtml.href: we had to resolve a path ourselves (normally done
by browser). It also inconvenient for the user when they see #zSoyz
without any explanation.

Initially the rationale for disallowing "/../" in FilterNormalizeUri was
to prevent a malicious user to inject "../../" in $local parameter which
is used in script path generation. We filter out scripts, so this case
does not apply to Markdown.

This CL disables the check by FitlerNormalizeUri.valueFilter and
simplifies MarkdownToHtml.href

Change-Id: I7b962966a25cb23b668365ec94be396c7c38a9c1