Update patch set 7
Patch Set 7:

Regarding the blame view, I made sure that the code content never went beyond the 80 char limit. Even if it does it won't require scrolling unless the lines are obscenely long. That is another view that I'm going to be working on next to get it in a better spot.

Patch-set: 7